fix: replace legacy code
